separation of some intersections (rect-rect, circle-circle, circle-point)

+++ gear/src/gear/collision/shapes/Circle.cpp
@@ -1,3 +1,3 @@
 #include "Circle.h"
 
-gear::Circle::Circle(const gear::Vector<float, 2> &position, float radius) : Shape(ShapeType::CIRCLE), position(position), radius(radius) {}
+gear::Circle::Circle(const gear::Vector<double, 2> &position, double radius) : Shape(ShapeType::CIRCLE), position(position), radius(radius) {}
