+++ gear/src/gear/collision/shapes/Point.cpp
@@ -1,4 +1,4 @@
 #include "Point.h"
 
-gear::Point::Point(const gear::Vector<float, 2> &position) : Shape(ShapeType::POINT), position(position) {}
+gear::Point::Point(const gear::Vector<double, 2> &position) : Shape(ShapeType::POINT), position(position) {}
