+++ gear/src/gear/collision/shapes/Point.h
@@ -7,9 +7,9 @@ _GEAR_START
 
 class Point : public Shape {
 public:
-  Vector<float, 2> position;
+  Vector<double, 2> position;
   
-  Point(const Vector<float, 2> &pos);
+  Point(const Vector<double, 2> &pos);
 
   //bool intersects(Shape *other) override;
   //void transform(TransformComponent *transform, Shape *transformed_Shape) override;
