+++ gear/src/gear/collision/shapes/Rect.h
@@ -8,10 +8,10 @@ _GEAR_START
 class Rect : public Shape
 {
 public:
-  Vector<float, 2> top_Left, bottom_Right;
+  Vector<double, 2> top_Left, bottom_Right;
 
   Rect(void);
-  Rect(const Vector<float, 2> &top_Left, const Vector<float, 2> &bottom_Right);
+  Rect(const Vector<double, 2> &top_Left, const Vector<double, 2> &bottom_Right);
 };
 
 _GEAR_END
